+++ common/src/main/java/com/pedro/common/av1/AV1Parser.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.pedro.rtmp.flv.video.av1
+package com.pedro.common.av1
 
 /**
  * Created by pedro on 8/12/23.
@@ -45,7 +45,7 @@ class AV1Parser {
 
   fun getObuType(header: Byte): ObuType {
     val value = ((header.toInt() and 0x7F) and 0xF8) ushr 3
-    return ObuType.values().firstOrNull { it.value == value } ?: ObuType.RESERVED
+    return ObuType.entries.firstOrNull { it.value == value } ?: ObuType.RESERVED
   }
 
   fun getObus(av1Data: ByteArray): List<Obu> {
